# Unroll first iteration of checked_ilog loop