@tspiteri 2024-04-23
Some benchmark results below:

Before #115913:

```
    num::int_log::u128_log_geometric        82.90µs/iter +/- 546.00ns
    num::int_log::u128_log_random            2.10ms/iter   +/- 3.56µs
    num::int_log::u128_log_random_small      1.43ms/iter   +/- 5.12µs
    num::int_log::u16_log_geometric        747.00ns/iter   +/- 0.00ns
    num::int_log::u16_log_random           633.83µs/iter   +/- 2.74µs
    num::int_log::u16_log_random_small     503.88µs/iter   +/- 4.12µs
    num::int_log::u32_log_geometric          4.35µs/iter  +/- 45.00ns
    num::int_log::u32_log_random           771.64µs/iter  +/- 34.28µs
    num::int_log::u32_log_random_small     206.21µs/iter   +/- 3.10µs
    num::int_log::u64_log_geometric         26.36µs/iter  +/- 66.00ns
    num::int_log::u64_log_random             1.09ms/iter  +/- 15.10µs
    num::int_log::u64_log_random_small     251.92µs/iter   +/- 5.14µs
    num::int_log::u8_log_geometric         180.00ns/iter   +/- 3.00ns
    num::int_log::u8_log_random            192.47µs/iter   +/- 2.52µs
    num::int_log::u8_log_random_small      192.22µs/iter   +/- 1.22µs
```

After #115913:

```
    num::int_log::u128_log_geometric        26.52µs/iter +/- 232.00ns
    num::int_log::u128_log_random            1.83ms/iter  +/- 10.74µs
    num::int_log::u128_log_random_small    936.77µs/iter   +/- 5.77µs
    num::int_log::u16_log_geometric        295.00ns/iter   +/- 8.00ns
    num::int_log::u16_log_random           465.97µs/iter   +/- 6.87µs
    num::int_log::u16_log_random_small     407.67µs/iter   +/- 8.13µs
    num::int_log::u32_log_geometric        800.00ns/iter  +/- 13.00ns
    num::int_log::u32_log_random           451.73µs/iter   +/- 2.07µs
    num::int_log::u32_log_random_small     468.66µs/iter   +/- 1.78µs
    num::int_log::u64_log_geometric          5.02µs/iter  +/- 34.00ns
    num::int_log::u64_log_random           593.26µs/iter   +/- 1.52µs
    num::int_log::u64_log_random_small     179.60µs/iter   +/- 2.88µs
    num::int_log::u8_log_geometric         129.00ns/iter   +/- 2.00ns
    num::int_log::u8_log_random            474.61µs/iter   +/- 2.00µs
    num::int_log::u8_log_random_small      474.85µs/iter   +/- 1.84µs
```

After this PR:

```
    num::int_log::u128_log_geometric        26.56µs/iter +/- 152.00ns
    num::int_log::u128_log_random            1.20ms/iter   +/- 4.21µs
    num::int_log::u128_log_random_small    623.08µs/iter   +/- 3.67µs
    num::int_log::u16_log_geometric        243.00ns/iter   +/- 1.00ns
    num::int_log::u16_log_random           194.60µs/iter   +/- 4.39µs
    num::int_log::u16_log_random_small     125.58µs/iter   +/- 4.08µs
    num::int_log::u32_log_geometric        733.00ns/iter  +/- 19.00ns
    num::int_log::u32_log_random           340.92µs/iter   +/- 2.27µs
    num::int_log::u32_log_random_small     143.57µs/iter   +/- 1.39µs
    num::int_log::u64_log_geometric          4.00µs/iter  +/- 26.00ns
    num::int_log::u64_log_random           465.61µs/iter   +/- 1.44µs
    num::int_log::u64_log_random_small     136.94µs/iter   +/- 2.69µs
    num::int_log::u8_log_geometric         107.00ns/iter   +/- 4.00ns
    num::int_log::u8_log_random            113.16µs/iter   +/- 1.59µs
    num::int_log::u8_log_random_small      113.86µs/iter   +/- 2.19µs
```