@tspiteri 2024-04-23
I'm not familiar with codegen tests (neither their purpose nor how to write them).