@workingjubilee 2024-04-23
when doing an optimization by hand that ideally the compiler could do implicitly, we prefer to have a codegen test or assembly test that validate the output is what we want. one day, this allows verifying the hand-tuned optimization is no longer needed. it also prevents the optimization from regressing in a later PR, so that your work is not undone by a change somewhere else. say, in a random LLVM update that will batch a lot of LLVM changes which will significantly affect many codegen instances and *can* regress some even though it is on average an improvement.

we use [FileCheck](https://llvm.org/docs/CommandGuide/FileCheck.html) for them. it's a somewhat esoteric syntax. I suggest looking at [another recent PR that came with a codegen test](https://github.com/rust-lang/rust/pull/124114/files#diff-97e6cb0a481ebbb10eec5352cb635897bc30b9f7acaff891e710a3a9f1a5fa26) to see a fairly "worked" example of one.