@workingjubilee 2024-04-23
that test in question uses `CHECK-DAG` but many tests are fine using `CHECK-NEXT` in sequence. `CHECK-DAG` just is less brittle when the ordering between two things can actually change without effect because what matters is they both come before a third.