@tspiteri 2024-04-23
I added a codegen test to confirm that when the value < base, there is a branch which returns without multiplication and division.